[REFACTOR] #75 - others 패키지 변경
- others -> utils로 패키지 이름 변경
- Utils object 기능 별로 분리
- GitAuth Uri 획득 함수 GithubApiInstance로 이동
