+++ app/src/main/java/com/example/android_repo_05/data/network/GithubApiInstance.kt
@@ -1,6 +1,8 @@
 package com.example.android_repo_05.data.network
 
-import com.example.android_repo_05.others.Constants
+import android.net.Uri
+import com.example.android_repo_05.BuildConfig
+import com.example.android_repo_05.utils.Constants
 import okhttp3.OkHttpClient
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
@@ -14,4 +16,12 @@ object GithubApiInstance {
         .client(client)
         .build()
         .create(GithubApi::class.java)
+
+    fun getGithubIdentityRequestUri(): Uri = Uri.Builder().scheme("https").authority("github.com")
+        .appendPath("login")
+        .appendPath("oauth")
+        .appendPath("authorize")
+        .appendQueryParameter("client_id", BuildConfig.GITHUB_CLIENT_ID)
+        .appendQueryParameter("scope", "repo,user,notifications")
+        .build()
 }
