+++ app/src/main/java/com/example/android_repo_05/utils/TimeUtils.kt
@@ -1,38 +1,12 @@
-package com.example.android_repo_05.others
+package com.example.android_repo_05.utils
 
-import android.content.Context
-import android.graphics.Color
-import android.net.Uri
-import com.example.android_repo_05.BuildConfig
-import com.example.android_repo_05.others.enums.TimeUnit
+import com.example.android_repo_05.utils.enums.TimeUnit
 import java.text.ParseException
 import java.text.SimpleDateFormat
 import java.util.*
-import kotlin.math.roundToInt
 
-object Utils {
+object TimeUtils {
     private val simpleDateFormat = SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ssX", Locale.KOREA)
-
-    val languageColorMap : MutableMap<String, Int> = mutableMapOf(
-        "Kotlin" to Color.rgb(169,123,255),
-        "Java" to Color.rgb(176,114,25),
-        "Python" to Color.rgb(53,114,165),
-        "JavaScript" to Color.rgb(241,224,90),
-        "C++" to Color.rgb(243,75,125)
-    )
-
-    fun getGithubIdentityRequestUri(): Uri = Uri.Builder().scheme("https").authority("github.com")
-        .appendPath("login")
-        .appendPath("oauth")
-        .appendPath("authorize")
-        .appendQueryParameter("client_id", BuildConfig.GITHUB_CLIENT_ID)
-        .appendQueryParameter("scope", "repo,user,notifications")
-        .build()
-
-    fun dpToPx(context: Context, dp: Int): Int {
-        val density = context.resources.displayMetrics.density
-        return (dp.toFloat() * density).roundToInt()
-    }
 
     fun calculateElapsedTime(updatedAt: String): String {
         try {
